# Conflicts:
#	src/main/java/org/posapp/model/Datastore.java

+++ src/main/java/org/posapp/model/Barang.java
@@ -15,7 +15,7 @@ public class Barang implements Serializable {
     private String nama;
     private String kategori;
     private int stok;
-    private int hargaJual;
-    private int hargaBeli;
+    private float hargaJual;
+    private float hargaBeli;
     private String pathGambar;
 }
